==> src/ipynb/Step_1-Preparation-Passengers.ipynb <==
# Function Codes

def dataCleaning(df, code=True, tips=False, orientation=True, formatIssues=True, missingValues=True, duplicateValues=True, outliers=True):
    """
    df: your dataframe

    code: A text template to note your observations as you go. Use the code snippets included in the output. copy-paste into vscode/notepad

    tips: Provides snippets of code to help you clean potential issues in your df. If you prefer this to code
    
    orientation: Provides information about the shape/objects of your data
    
    formatIssues: Provides detailed information on each column to help identify format issues
    
    missingValues: Provides information on missing values
    
    duplicateValues: Provides information on duplicate values
    
    outliers: Provides information on outliers
    """
    import pandas as pd
    import numpy as np
    import matplotlib.pyplot as plt
    import seaborn as sns

    if code==True:
        print("### CLEANING CODE:")
        print("df = dfX #Change to your df's name")
        print()
        print("#### Change column value:")
        print()
        print()
        print("#### Drop entire column:")
        print()
        print()
        print("#### Change column type:")
        print()
        print()
        print("#### Change column name:")
        print()
        print()
        print("#### Handle missing values:")
        print()
        print()
        print("#### Handle duplicate values:")
        print("# df.drop_duplicates(inplace=True) # drop ALL duplicate rows")
        print()
        print("#### Drop outliers:")
        print()
        print()
        print("#### Other observations / further investigations:")
        print("#")
        print("#")
        print("#")
        print()
        print("df.head() #Final Review")
        print("# dfX = df #Change to your df's name")
        print()
        print("=========================================")
    
    if orientation==True:
        print("ORIENTATION")
        print(df.info())
        print("=========================================")
        print()
        
    
    if formatIssues==True:
        print("FORMAT ISSUES")
        print()
        for col in df.columns:
            if df[col].dtype == 'object' or df[col].dtype == 'int64' or df[col].dtype == 'float64' or df[col].dtype == 'datetime64':
            #if df[col].dtype == 'float64':

                print("df.rename(columns={'" + col + "': ''}, inplace=True)", "#rename column")
                print("df['" + col + "'] = df['" + col + "'].replace('old_value', 'new_value')")
                print("df['" + col + "'] = df['" + col + "'].astype('new_type') # new_type can be int64, float64, object, category, datetime64")
                print("df.drop('" + col + "', axis=1, inplace=True)")                
                pd.set_option('display.max_rows', None)
                print(df.groupby(col, sort=True).size())
                pd.reset_option('display.max_rows')
                #display the dtypes of the column
                print("Current Column DType: ", df[col].dtype, "     Do not compare with above. This one will always return int64 as it's the dtype of the count")                
                print("df['" + col + "'] = df['" + col + "'].astype('new_type') # new_type can be int64, float64, object, category, datetime64")
                print()
            #else:
            #    print(col)
            #    print(df[col].describe())
            #    print()

        if tips==True:
            print("TIPS")
            print("To make a correction to a column, use the following syntax:")
            print("df['A'] = df['A'].apply(lambda x: x.replace('old_value', 'new_value'))")
            print()
            print("To change a column data type, use the following syntax:")
            print("df['A'] = pd.to_datetime(df['A']) # for datetime")
            print("df['A'] = df['A'].astype('int64') # for integers")
            print("df['A'] = df['A'].astype('float64') # for floats")
            print("df['A'] = df['A'].astype('category') # for categorical")
            print("df['A'] = df['A'].astype('object') # for object")
            print()
        print("=========================================")
        print()

    if missingValues==True:
        print("MISSING VALUES")
        print()
        for col in df.columns:
            if df[col].isnull().sum() > 0:
                print(col, ":", df[col].isnull().sum(), " missing values")
                print("df.dropna(subset=['" + col + "'], inplace=True)")
                print("df['" + col + "'].fillna(df['" + col + "'].mean(), inplace=True) #fill NA entries with the mean")
                print("df['" + col + "'].fillna(0, inplace=True) # fill NA entries with a single value, such as zero")
                print()
                print(df.loc[df[col].isnull()].head())
                print()
            else:
                print(col, ": No missing values")
                print()
                                    
        if tips==True:
            print()
            print("TIPS")
            print("You can drop rows with missing values using one of the following code:")
            print("df.dropna(subset=['col'], inplace=True) #For a single column")
            print("df.dropna(inplace=True) #For all columns")
            print()
            print("You can fill rows with missing values using one of the following code:")
            print("df['col'].fillna(df['col'].mean(), inplace=True) #fill NA entries with the mean")
            print("df['col'].fillna(0, inplace=True) # fill NA entries with a single value, such as zero")
            print("df['col'].fillna(method='ffill') # forward-fill to propagate the previous value forward")
            print("df['col'].fillna(method='bfill' # back-fill to propagate the next values backward)")
            print()
            print("To view them:")
            print("df.loc[df[col].isnull()].head()")
            print()
        print("=========================================")
        print()

    if duplicateValues==True:
        print("DUPLICATE VALUES")
        print()
        print(df[df.duplicated()].head())
        print()

        if tips==True:
            print("TIPS")
            print("You can drop duplicate rows using the following code:")
            print("df.drop_duplicates(inplace=True)")
            print("df.drop_duplicates(subset=['col'], inplace=True) #For a single column")
            print()
            print("To view them:")
            print("df[df.duplicated()].head()")
            print()
    
        print("=========================================")
        print()

    if outliers==True:
        print("OUTLIERS")
        print()
        for col in df.columns:
            if df[col].dtype == 'int64' or df[col].dtype == 'float64':
                print(col)
                print("-----")
                print("Outlier(s):")
                print("Below ", df[col].mean() - 3*df[col].std(), " -> ", df[df[col] < df[col].mean() - 3*df[col].std()].shape[0], " low outlier(s)")
                print("Above ", df[col].mean() + 3*df[col].std(), " -> ", df[df[col] > df[col].mean() + 3*df[col].std()].shape[0], " high outlier(s)")
                low = df[col].mean() - 3*df[col].std()
                high = df[col].mean() + 3*df[col].std()
                print("df = df[(df['" + col + "'] > " + str(low) + ") & (df['" + col + "'] < " + str(high) + ")]")
                print()
                print(df[col].describe())
                print()
                print("Boxplot")
                sns.boxplot(df[col])
                plt.show()
                print()
                print("Histogram")
                sns.histplot(df[col])
                plt.show()
                print("=========================================")
                print()

        if tips==True:
            print("TIPS")
            print("You can drop outliers using the following code:")
            print("df = df[(df['column'] > lower_bound) & (df['column'] < upper_bound)]")
            print()

# Import

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
pd.set_option("display.max_columns", 120)

flights_sample = pd.read_csv("../../data/raw/flights_sample+test.csv")

# passengers = pd.read_csv("../../data/raw/passengers_w_departuresPerformed(29Nov).csv") 
## In the first step noticed this was not properly done (see Take 1 - Archived) so made the new one below.
## Uncomment above and switch the cells in Take 1 archived to code for further details

passengers = pd.read_csv("../../data/raw/passengers_w_departuresPerformed_groupedbyMonth(29Nov).csv", index_col=None)

passengers:

This CSV was produced using the following SQL Query on the passengers table:

SELECT CONCAT(year, '_', month, '-', unique_carrier, '-', origin, '-', dest) AS routeId
	, payload / departures_performed AS averagePayload_lbs
	, freight / departures_performed AS averageFreight_lbs
	, mail / departures_performed AS averageMail_lbs
	, seats / departures_performed AS availableSeats
	, passengers / departures_performed AS averagePassengers
	--, passengers / seats * 100 AS averageSeatsoccupied_perc
	, aircraft_group AS aircraftGroup
	, aircraft_type AS aircraftType
	, aircraft_config AS aircraftConfiguration
	, distance_group AS distanceInterval_x500mi
	, class AS serviceClass
FROM passengers
WHERE departures_performed > 0

# Deciding on the Route ID column (will be used to merge with flights)

## Take 1 (dropped)

The passengers table is huge, it likely has information on routes that we don't need. Let's slim it down

passengers.shape

###Create the route_ID column for flights_sample
flights_sample['routeId_op'] = flights_sample['Flight Year'].astype(str) + '_' + flights_sample['Flight Month'].astype(str) + '-' + flights_sample['Operator - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']
flights_sample['routeId_mkt'] = flights_sample['Flight Year'].astype(str) + '_' + flights_sample['Flight Month'].astype(str) + '-' + flights_sample['Marketer - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']

Made 2x version, as not sure if the passengers' table is referring to the Operator or the Marketer.. we'll test both

## List the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy arrai
flight_op = pd.DataFrame()
flight_mkt = pd.DataFrame()
unique_passengers = pd.DataFrame() 

flight_op['routeId_op'] = flights_sample['routeId_op'].unique()
flight_mkt['routeId_mkt'] = flights_sample['routeId_mkt'].unique()
unique_passengers['routeid'] = passengers['routeid'].unique()

## Count the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy array
print("Flight Operator")
print("Total:", len(flight_op))
print("In passenger:", flight_op['routeId_op'].isin(unique_passengers['routeid']).sum())
print()
print("Flight Marketer")
print("Total:", len(flight_mkt))
print("In passenger:", flight_mkt['routeId_mkt'].isin(unique_passengers['routeid']).sum())
print()
print("Passengers Table") #Double tab but just to be sure..
print("Total:", len(unique_passengers))
print("In Marketer:", unique_passengers['routeid'].isin(flight_mkt['routeId_mkt']).sum())
print("In Operator:", unique_passengers['routeid'].isin(flight_op['routeId_op']).sum())

Ok so the Flight Operator has more results, let's see the route that are not in the passengers table

print(flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].shape)
flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].head(10)

This just made me notice something.. we don't have the data for 2020.. so we need to consolidate this information on a monthly thing. We'll need to reformat our route_ID to MM-carrier-origin-dest and then group the info of every year to either an average or a min/max where appropriate..

## Take 2 (dropped)

###Create the route_ID column for flights_sample
flights_sample['routeId_op'] = flights_sample['Flight Month'].astype(str) + '-' + flights_sample['Operator - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']
flights_sample['routeId_mkt'] = flights_sample['Flight Month'].astype(str) + '-' + flights_sample['Marketer - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']

## List the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy arrai
flight_op = pd.DataFrame()
flight_mkt = pd.DataFrame()
unique_passengers = pd.DataFrame() 

flight_op['routeId_op'] = flights_sample['routeId_op'].unique()
flight_mkt['routeId_mkt'] = flights_sample['routeId_mkt'].unique()
unique_passengers['routeid'] = passengers['routeid'].unique()

## Count the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy array
print("Flight Operator")
print("Total:", len(flight_op))
print("In passenger:", flight_op['routeId_op'].isin(unique_passengers['routeid']).sum())
print()
print("Flight Marketer")
print("Total:", len(flight_mkt))
print("In passenger:", flight_mkt['routeId_mkt'].isin(unique_passengers['routeid']).sum())
print()
print("Passengers Table") #Double tab but just to be sure..
print("Total:", len(unique_passengers))
print("In Marketer:", unique_passengers['routeid'].isin(flight_mkt['routeId_mkt']).sum())
print("In Operator:", unique_passengers['routeid'].isin(flight_op['routeId_op']).sum())

Ok so the Flight Operator has more results, let's see the route that are not in the passengers table

print(flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].shape)
flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].head(10)

Ok, it looks like a fair amount is from our test... let's see...

count = flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])]
count['Flight Year'].value_counts()

This is disappointing.. this probably means that we have new routes in our test..

## Take 3 - Month-depart-arrival (dropped)

###Create the route_ID column for flights_sample
flights_sample['routeId_v3'] = flights_sample['Flight Month'].astype(str) + '-' +  flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']
## We just need one in this case

# Remove the carrier from the route ID column
unique_passengers = pd.DataFrame()
unique_passengers['routeid'] = passengers['routeid'].unique()

unique_passengers['routeid_month'] = unique_passengers['routeid'].str[:-11]
unique_passengers['routeid_route'] = unique_passengers['routeid'].str[-8:]

unique_passengers['routeid'] = unique_passengers['routeid_month'] + unique_passengers['routeid_route']

Let's see how this performs:

## List the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy arrai
flight_op = pd.DataFrame()

flight_op['routeId_op'] = flights_sample['routeId_v3'].unique()

## Count the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy array
print("Flights")
print("Total:", len(flight_op))
print("In passenger:", flight_op['routeId_op'].isin(unique_passengers['routeid']).sum())
print()
print("Passengers Table") #Double tab but just to be sure..
print("Total:", len(unique_passengers))
print("In Flight:", unique_passengers['routeid'].isin(flight_op['routeId_op']).sum())

This seems to perform better. Let's confirm

print(flights_sample[~flights_sample['routeId_v3'].isin(unique_passengers['routeid'])].shape)
flights_sample[~flights_sample['routeId_v3'].isin(unique_passengers['routeid'])].head(10)

Still some missing from Jan 2020, damn!

count = flights_sample[~flights_sample['routeId_v3'].isin(unique_passengers['routeid'])]
count['Flight Year'].value_counts()

count['Flight Month'].value_counts()

flights_sample[(flights_sample['Flight Year'] == 2020) & (flights_sample['Flight Month'] == 1)].shape

6741/660556*100

So we account for about 99% of the routes in January. It is better, but there is room for improvement. Also not a fan that we group all carriers together.. let's look at the other methods.  

## Take 4 - carrier-depart-arrival

###Create the route_ID column for flights_sample
flights_sample['routeId_op'] = flights_sample['Operator - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']
flights_sample['routeId_mkt'] = flights_sample['Marketer - Unique Carrier Code'] + '-' + flights_sample['Origin Airport (IATA Code)'] + '-' + flights_sample['Destination Airport (IATA Code)']

# Remove the carrier from the route ID column
unique_passengers = pd.DataFrame()

unique_passengers['routeid'] = passengers['routeid'].unique()
unique_passengers['routeid'] = unique_passengers['routeid'].str[-10:]

Let's see how this performs:

## List the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy arrai
flight_op = pd.DataFrame()
flight_mkt = pd.DataFrame()

flight_op['routeId_op'] = flights_sample['routeId_op'].unique()
flight_mkt['routeId_mkt'] = flights_sample['routeId_mkt'].unique()

## Count the number of unique flights_sample['routeId_op'] values that are in the passengers['routeid'] column using the isin() method, ensuring the values are stored in a df, not a numpy array
print("Flight Operator")
print("Total:", len(flight_op))
print("In passenger:", flight_op['routeId_op'].isin(unique_passengers['routeid']).sum())
print()
print("Flight Marketer")
print("Total:", len(flight_mkt))
print("In passenger:", flight_mkt['routeId_mkt'].isin(unique_passengers['routeid']).sum())
print()
print("Passengers Table") #Double tab but just to be sure..
print("Total:", len(unique_passengers))
print("In Marketer:", unique_passengers['routeid'].isin(flight_mkt['routeId_mkt']).sum())
print("In Operator:", unique_passengers['routeid'].isin(flight_op['routeId_op']).sum())

This seems to perform better, and again Flight Operator is more commonly used. Let's confirm

print(flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].shape)
flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])].head(10)

Still some missing from Jan 2020, but we actually seem to have less this time. Let's confirm

count = flights_sample[~flights_sample['routeId_op'].isin(unique_passengers['routeid'])]
count['Flight Year'].value_counts()

count['Flight Month'].value_counts()

it is better, but there is room for improvement, let's look at the other methods. Also not a fan that we group all carriers together.. 

flights_sample[(flights_sample['Flight Year'] == 2020) & (flights_sample['Flight Month'] == 1)].shape

993/660556*100

Okay, the proportion is almost negligible at this point We account for 99.85% of the routes

# Grouping the values

#Lets see the duplicates

passengers['routeid'] = passengers['routeid'].str[-10:]


print("Total rows:", passengers['routeid'].count())
print("Unique values:", passengers['routeid'].nunique())
print("Duplicate values:", passengers['routeid'].duplicated().sum())

So each rows have about 5-6x values.. we'll group them together

Before we group them, let's calculate the Freight ratio, the Mail Ratio as well as the amount of SeatOccupied, it will be more accurate than calculating once they are merged together

# The 0s seem the be throwing off our calculations.. we'll switch them to NaNs
passengers['availableseats'] = passengers['availableseats'].replace(0, np.nan)
passengers['averagepassengers'] = passengers['averagepassengers'].replace(0, np.nan)
passengers['averagepayload_lbs'] = passengers['averagepayload_lbs'].replace(0, np.nan)
passengers['averagefreight_lbs'] = passengers['averagefreight_lbs'].replace(0, np.nan)
passengers['averagemail_lbs'] = passengers['averagemail_lbs'].replace(0, np.nan)

passengers['Proportion of freight to the payload'] = passengers['averagefreight_lbs'] / passengers['averagepayload_lbs']
passengers['Proportion of mail to the payload'] = passengers['averagemail_lbs'] / passengers['averagepayload_lbs']
passengers['Proportion of filled seats'] = passengers['averagepassengers'] / passengers['availableseats']

We will group by routeid and avg the values for averagepayload_lbs, averagefreight_lbs, averagemail_lbs, availableseats, averagepassengers, distanceinterval_x500mi 

As well, aircraftgroup, aircrafttype, aircraftconfiguration, and serviceclass are all categorical variables, so we'll take the most common value for each group and return UNK if we don't have the info

We're putting the grouping results into grouped_passengers, below:

#Switching the NANs to 'UNK'
passengers['serviceclass'].fillna('UNK', inplace=True)
passengers['aircraftgroup'].fillna('UNK', inplace=True)
passengers['aircrafttype'].fillna('UNK', inplace=True)
passengers['aircraftconfiguration'].fillna('UNK', inplace=True)

grouped_passengers = passengers.groupby(['routeid']).agg({'Proportion of freight to the payload': 'mean', 'Proportion of mail to the payload': 'mean', 'Proportion of filled seats': 'mean', 'averagepayload_lbs': 'mean', 'availableseats': 'mean', 'distanceinterval_x500mi': 'mean', 'aircraftgroup': lambda x: x.value_counts().index[1] if x.value_counts().index[0] == 'UNK' and len(x.value_counts()) > 1 else x.value_counts().index[0], 'aircrafttype': lambda x: x.value_counts().index[1] if x.value_counts().index[0] == 'UNK' and len(x.value_counts()) > 1 else x.value_counts().index[0], 'aircraftconfiguration': lambda x: x.value_counts().index[1] if x.value_counts().index[0] == 'UNK' and len(x.value_counts()) > 1 else x.value_counts().index[0], 'serviceclass': lambda x: x.value_counts().index[1] if x.value_counts().index[0] == 'UNK' and len(x.value_counts()) > 1 else x.value_counts().index[0]}).reset_index()

print(passengers.shape)
print(grouped_passengers.shape)
grouped_passengers.sample(10)

And we're matching our unique values from above.. TRIPLE BAM!

# Export to CSV

grouped_passengers.to_csv("../../data/processed/flights_enrichment_avgpayload_passengers.csv")
