==> src/ipynb/Step_3-Modeling.ipynb <==
## Import Dataset & Librairies

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.preprocessing import PolynomialFeatures
import xgboost as xgb
from xgboost import XGBClassifier
from xgboost import XGBRegressor

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error

flights_sample_train= pd.read_csv('../../data/processed/Enriched-flights_sample_train.csv') # Used for training model
flights_sample_test= pd.read_csv('../../data/processed/Enriched-flights_sample_test.csv') # Used for testing model

flights_test= pd.read_csv('../../data/processed/Enriched-flights_BIGtest(2dec).csv') # Simulation of the test data set. Full random 200K
flights_final_test= pd.read_csv('../../data/processed/Enriched-flights_test.csv') # Mid_term Project Test dataset

# Preferred model:

## XGBoost Regressor w/ GridSearch  RMSE: 20.38, 45.76, 48.39

### Tune Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor
from sklearn.model_selection import GridSearchCV

model_param_grid = {'learning_rate': [1,0.1,0.01,0.001],
                  'n_estimators': [10,25,50,100,200,300,400,500],
                  'subsample': [0.1,0.25,0.5,0.75,0.9],
                  'reg_lambda':[1,2,3,4,5],
                  'gamma':[1,2,3,4,5],
                  'max_depth':[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15]}

model = xgb.XGBRegressor()

grid_mse = GridSearchCV(estimator=model, param_grid=model_param_grid, scoring='neg_mean_squared_error',cv=4, verbose=1)

# model.fit(X_train, y_train)
grid_mse.fit(X_train, y_train)

print("Best parameters found: ", grid_mse.best_params_)
print("Lowest RMSE found: ", np.sqrt(np.abs(grid_mse.best_score_)))

Results were:
learning rate = .01
n_estimators = 400
reg_lambda = 1
gamma = 1
subsample = 0.5
max_depth = 11

Re-running above takes many hours.

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor
from sklearn.model_selection import GridSearchCV

model = xgb.XGBRegressor(learning_rate=0.01, n_estimators=400, subsample=0.5, max_depth=11, gamma=1, reg_lambda=1)
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

print("MSE:",mean_squared_error(y_test, y_pred))

print("train score:", model.score(X_train, y_train))
print("test score:", model.score(X_test, y_test))

### Save model into a pickle

import pickle
pickle.dump(model, open('../pickles/XGBoost+model.pkl', 'wb'))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
#flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

print("MSE:",mean_squared_error(y, y_pred))
print("Score:", model.score(X, y))

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
# del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']


## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

print("MSE:",mean_squared_error(y, y_pred))
print("Score:", model.score(X, y))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/XGBoostReg.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/XGBoostReg_tuned.csv', index=False)

# Other models

## XGBoost Regressor RMSE: 20.93, 45.98, 48.64

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor

model = XGBRegressor()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))
print("MSE:",mean_squared_error(y_test, y_pred))

print("train score:", model.score(X_train, y_train))
print("test score:", model.score(X_test, y_test))

overfitting.. need to tune

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']


## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/XGBoostReg.csv', index=False)

### Final Test (Jan2020)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/XGBoostReg.csv', index=False)

## XGBoost Regressor - What if we added more data? RMSE (40.00, 46.05)

Let's add our 200K to our training data and see if we can better predict the 40K

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
#flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

flights_sample_combined_train = pd.concat([flights_sample_train, flights_test], axis=0)

### Train the model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_combined_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_combined_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor

model = XGBRegressor()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

So this is not really helping our initial value, but what about the 40K?

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_combined_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_combined_train.columns))

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Nope.. this got worse..

### Not a tuning per say, but what is the accuracy score of XGBoost Reg?

## XGBoost Regressor - Tuning the features, not the model(dropped)

### Let's try keeping only our correlated features  RMSE: 21.20

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
y=flights_sample_train['Arrival Delay (minutes)']

#Let's try keeping only our correlated features
X = X[X.columns[(X.corrwith(y) > 0.01).tolist()]]

X=X.fillna(0)
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor

model = XGBRegressor()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

Well that's not really doing any better...

### How about Min/Max Scaler?  RMSE: 20.93

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
y=flights_sample_train['Arrival Delay (minutes)']

X=X.fillna(0)
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# Scaling the input features using MinMaxScaler
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
scaler.fit(X)
X_scaled = scaler.transform(X)

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBRegressor

model = XGBRegressor()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

Not doing anything

## Random Forest Regressor   RMSE: 20.681228, 45.845254, 48.44

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
from sklearn.ensemble import RandomForestRegressor

model = RandomForestRegressor(n_estimators=100, random_state=1)
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))


print("train score:", model.score(X_train, y_train))
print("test score:", model.score(X_test, y_test))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']

## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/RandomForestReg.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/RandomForestReg.csv', index=False)

## Random Forest Classifier   RMSE: 24.62, 49.42, 51.80  Submitted: Yes

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
from sklearn.ensemble import RandomForestClassifier

model = RandomForestClassifier()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']

## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/RandomForestCl.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/RandomForestCl.csv', index=False)

## Linear Regression  RMSE: 994140.93, 2435111.90, 2037559.82  Submitted: No

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
from sklearn.linear_model import LinearRegression

model = LinearRegression()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']

## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/LinReg.csv', index=False)

### Final Test (Jan2020) - Not submitting this one results are too bad

## SVM RMSE:21.86, 47.77, ?  Export:No...........Takes too long!

SVM is extremely slow... so far we haven't been able to finish running the models

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.2
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import sklearn
from sklearn import svm

# model = svm.SVR()
# model.fit(X_train, y_train)
model = pickle.load(open('../SVMmodel.pkl', 'rb')) # Using the pickle'd version of the trained model

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
from sklearn.metrics import mean_squared_error
from math import sqrt

rms = sqrt(mean_squared_error(y_test, y_pred))
print(rms)

### Save model into a pickle

import pickle
pickle.dump(model, open('../pickles/SVMmodel.pkl', 'wb'))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
#flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

### Second Test (200K) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']


## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

model = pickle.load(open('../pickles/SVMmodel.pkl', 'rb'))

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/SVM.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']


## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values


model = pickle.load(open('../pickles/SVMmodel.pkl', 'rb'))

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/SVM.csv', index=False)

## XG Boost Classifier - Error at training

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import xgboost as xgb
from xgboost import XGBClassifier

model = XGBClassifier()
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']

## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/XGBoostReg.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/XGBoostReg.csv', index=False)

## OLS - Error at training

### Train Model

# Let's set X and y to our flights_sample_train data and 'Arriaval Delay (minutes)' column
X=flights_sample_train.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_train['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# split into train and test sets
seed = 1
test_size = 0.8
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed)

# fit model to training data
import statsmodels.api as sm

model = sm.OLS
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

### First Test (40K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_sample_test.columns))
print(set(flights_sample_test.columns) - set(flights_sample_train.columns))

## Adding the missing column to the test
flights_sample_test['Aircraft type2_887.0'] = 0

## Not let's put our columns in the same order:
flights_sample_test = flights_sample_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_sample_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_sample_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

Well that's not good..

### Second Test (200K)

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_test.columns))
print()
print(set(flights_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set)

## Remove additional columns from the test (2nd set)
#del flights_test['Aircraft type2_837.0']
del flights_test['origin']
del flights_test['mkt_carrier']
del flights_test['dest']
del flights_test['mkt_carrier_fl_num']
del flights_test['fl_date']

## Now let's put our columns in the same order:
flights_test = flights_test[flights_sample_train.columns]

#### Now run the test

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)
y=flights_test['Arrival Delay (minutes)']
y=y.fillna(0)

# convert to numpy array
X=X.values
y=y.values

# make predictions for test data
y_pred = model.predict(X)

# evaluate predictions
rmse = np.sqrt(mean_squared_error(y, y_pred))
print("RMSE: %f" % (rmse))

#### Export 2nd Test Results to CSV (Final QC)

# output y and y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)
y_test_df=pd.DataFrame(y)

# merge the two dataframes
y_pred_df['y_test']=y_test_df

# rename the columns
y_pred_df.columns=['Model Prediction', 'Actual Value']

#export to csv
y_pred_df.to_csv('../../src/modules/2nd Test Predictions/OLS.csv', index=False)

### Final Test (Jan2020) - Incomplete

#### Let's start by making sure both have the same columns and are in the same order

print(set(flights_sample_train.columns) - set(flights_final_test.columns))
print()
print(set(flights_final_test.columns) - set(flights_sample_train.columns))

## Add missing columns to the test (1st set) ... Most of those will get dropped right after anyway.. it's really just so the arrays match
flights_final_test['cancelled'] = 0
flights_final_test['Weather Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_9K'] = 0
flights_final_test['diverted'] = 0
flights_final_test['Carrier Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_KS'] = 0
flights_final_test['National Air System Delay (minutes)'] = 0
flights_final_test['Operator - Unique Carrier Code2_VX'] = 0
flights_final_test['Aircraft type2_432.0'] = 0
flights_final_test['Security Delay (minutes)'] = 0
flights_final_test['Departure Delay (minutes)'] = 0
flights_final_test['cancellation_code'] = 0
flights_final_test['Arrival Delay (minutes)'] = 0
flights_final_test['Late Aircraft Delay (minutes)'] = 0

## Remove additional columns from the test (2nd set)



## Now let's put our columns in the same order:
flights_final_test = flights_final_test[flights_sample_train.columns]

#### Now run the test.. no score for this one, we're just predicting!

# Now let's try to predict the arrival delay for the flights_sample_test data
X=flights_final_test.drop(['Marketer - Unique Carrier Code', 'Operator - Unique Carrier Code', 'Tail Number', 'Flight Number', 'Aircraft group', 'Aircraft type', 'Aircraft configuration', 'Service class', 'Distance (miles)', 'Origin Airport (IATA Code)', 'Airport Type_origin', 'Scheduled Departure Time (local time)', 'Destination Airport (IATA Code)', 'Airport Type_destination', 'Scheduled Arrival Time (local time)', 'Departure Delay (minutes)', 'Arrival Delay (minutes)', 'Carrier Delay (minutes)', 'Weather Delay (minutes)', 'National Air System Delay (minutes)', 'Security Delay (minutes)', 'Late Aircraft Delay (minutes)', 'cancelled', 'cancellation_code', 'diverted'], axis=1)
X=X.fillna(0)

# convert to numpy array
X=X.values

# make predictions for test data
y_pred = model.predict(X)

### Export Final_Test Results to CSV

# output y pred into a dataframe
y_pred_df=pd.DataFrame(y_pred)

# Rename the columns
y_pred_df.columns=['predicted_delay']

# Now, we have to format y_pred_df to match the format of the sample submission file
## Produce the missing columns
flights_test['fl_date'] = flights_test['Flight Year'].astype(str) + '-' + flights_test['Flight Month'].astype(str).str.zfill(2) + '-' + flights_test['Flight Day'].astype(str).str.zfill(2)
flights_test['mkt_carrier'] = flights_test['Marketer - Unique Carrier Code']
flights_test['mkt_carrier_fl_num'] = flights_test['Flight Number']
flights_test['origin'] = flights_test['Origin Airport (IATA Code)']
flights_test['dest'] = flights_test['Destination Airport (IATA Code)']

## Add these columns to y_pred_df
y_pred_df['fl_date'] = flights_test['fl_date']
y_pred_df['mkt_carrier'] = flights_test['mkt_carrier']
y_pred_df['mkt_carrier_fl_num'] = flights_test['mkt_carrier_fl_num']
y_pred_df['origin'] = flights_test['origin']
y_pred_df['dest'] = flights_test['dest']

## Reorder the columns to match the sample submission file
y_pred_df = y_pred_df[['fl_date', 'mkt_carrier', 'mkt_carrier_fl_num', 'origin', 'dest', 'predicted_delay']]

# Export to csv
y_pred_df.to_csv('../../output/OLS.csv', index=False)

##
